Refactor double to BigDecimal

+++ BSCIdeas_intro/src/model/actions/AddPayment.java
@@ -1,15 +1,16 @@
 package model.actions;
 
+import java.math.BigDecimal;
 import model.Currency;
 import model.Payment;
 import model.Payments;
 
 public class AddPayment implements Action {
 
-    private final double amount;
+    private final BigDecimal amount;
     private final Currency currency;
 
-    public AddPayment(double amount, Currency currency) {
+    public AddPayment(BigDecimal amount, Currency currency) {
         this.amount = amount;
         this.currency = currency;
     }
